+++ CHANGELOG.md
@@ -46,9 +46,9 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 * Reactor command names can now be registered with the `CommandName` attribute.
 * Reactor command descriptions can now be registered with the `CommandDescription` attribute.
 * Reactor command arguments can now be registered with the `Arguments` attribute.
+* Added new `Cursor` class that makes it easier to control the CLI cursor.
 * Added a signal handler to make it easier to handle async process control signals.
 * Added a spinner CLI output helper.
-* It is now possible to hide and show the cursor in the CLI.
 * New and improved look of rendered CLI tables.
 
 #### Changes